refactor: modularize into submodules
- Split `peyepeline.py` into dedicated submodules: `signals`, `io`, `mapping`, `events`, and `data_collection`.
- Ensure key functionality is correctly routed through the API.
