+++ preprocessing/__init__.py
@@ -1,0 +1,7 @@
+"""Pipeline for MultiplEYE data preprocessing"""
+
+from .api import *  # Brings all functions from API into the top-level preprocessing namespace
+
+# Functionality made available with absolute imports
+from .utils import __all__
+from .data_collection import __all__
