+++ preprocessing/data_collection/__init__.py
@@ -1,0 +1,7 @@
+"""Data collection submodule of the preprocessing module."""
+
+from .multipleye_data_collection import MultipleyeDataCollection
+
+__all__ = [
+    "MultipleyeDataCollection",
+]
