+++ preprocessing/mapping/__init__.py
@@ -1,0 +1,1 @@
+"""Mapping submodule of the preprocessing module."""
